+++ scripts/handlePostSummon.js
@@ -1,6 +1,6 @@
 import { SOURCES, EFFECTS, MODULE_ID } from "./const.js";
 
-export async function handlePostSummon(itemUUID, actorUUID) {
+export async function handlePostSummon(itemUUID, actorUUID, summonerToken) {
     switch (itemUUID) {
         case SOURCES.COMMANDER.PLANT_BANNER:
             setTimeout(function () {
@@ -16,6 +16,36 @@ export async function handlePostSummon(itemUUID, actorUUID) {
 
             }, 1500) // DO this after 0.5 seconds to hopefully fix the no stuff applied yet issue
             break;
+        case SOURCES.MISC.WOODEN_DOUBLE:
+            if (!summonerToken) return;
+            const mvmntLocation = await Sequencer.Crosshair.show(
+                {
+                    location: {
+                        obj: summonerToken,
+                        showRange: true
+                    },
+                    label: {
+                        text: game.i18n.localize("pf2e-summons-assistant.display-text.wooden-double.step")
+                    },
+                    icon: {
+                        texture: summonerToken.document.texture.src
+                    },
+                    snap: {
+                        position: summonerToken.document.width % 2 === 1
+                            ? CONST.GRID_SNAPPING_MODES.CENTER
+                            : CONST.GRID_SNAPPING_MODES.VERTEX,
+                    },
+                    gridHighlight: true,
+                }
+            )
+
+            await new Sequence()
+                .animation()
+                .on(summonerToken)
+                .moveTowards(mvmntLocation, { relativeToCenter: true })
+                .play()
+            break;
+        //TO do set 
         default:
             break;
     }